requirements.txt
Flask>=2.3
requests>=2.25
python-dotenv>=1.0